+++ assets/javascripts/kitten/components/form/dropdown-select/combobox.js
@@ -47,6 +47,7 @@ export const DropdownCombobox = ({
   openOnLoad,
   uniqLabelOnSearch,
   menuZIndex,
+  className,
 }) => {
   const [flattenedOptions, setFlattenedOptions] = useState([])
   const [filteredOptions, setFilteredOptions] = useState([])
@@ -80,7 +81,10 @@ export const DropdownCombobox = ({
   }
 
   const onIsOpenChange = changes => {
-    if (changes.isOpen) return onMenuOpen({ changes })
+    if (changes.isOpen) {
+      flattenedOptions && setFilteredOptions(flattenedOptions)
+      return onMenuOpen({ changes })
+    }
 
     return onMenuClose({
       changes,
@@ -146,6 +150,7 @@ export const DropdownCombobox = ({
         'k-Form-Dropdown',
         `k-Form-Dropdown--${variant}`,
         `k-Form-Dropdown--${size}`,
+        className,
         {
           'k-Form-Dropdown--isOpen': isOpen > 0,
           'k-Form-Dropdown--error': error,