+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix `ToggleTip` first coordinates update.
 - Feature: Add all sizes to `DropdownSelect`.
 - Fix: Fix styles on `DropdownSelect`.
+- Fix: On `DropdownSelect combobox`, display full list of options on menu open.
+- Fix: Propagate `className` on `DropdownSelect` and `DropdownSelect combobox`.
+- Fix: Fix styles on `DropdownSelect`.
 
 ## [2.125.0] - 2021-03-15